removed old Optimizer class, refactored new atomic optimizers
