+++ doc/optimize.rst
@@ -1,3 +1,3 @@
-.. automodule:: openqml.optimizer
+.. automodule:: openqml.optimize
    :members:
    :private-members:
